microservices/backend-service/requirements.txt
Django==3.1.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
requests==2.25.1